moved generic mailing logic outside of scheduler
