+++ test/api/user_password_reset.test.js
@@ -40,7 +40,7 @@ describe('password reset', () => {
         stub.returns(Promise.resolve());
 
         const sendEmail = proxyquire('../../src/application/sendResetPasswordEmail', {
-            '../scheduler/sendResonatorEmail': {
+            '../mailing': {
                 default: stub
             },
             '../../application/getUow': {
